+++ packages/stylist/src/lib.rs
@@ -1,7 +1,9 @@
 #![deny(clippy::all)]
+#![deny(missing_debug_implementations)]
 #![deny(unsafe_code)]
 #![deny(non_snake_case)]
 #![cfg_attr(documenting, feature(doc_cfg))]
+#![cfg_attr(any(releasing, not(debug_assertions)), deny(dead_code, unused_imports))]
 
 //! Stylist is a CSS-in-Rust styling solution for WebAssembly Applications.
 //!